### 2019-07-17 #60 @moorepants
@labarba I can edit this one if you'd like and I have some people in mind that could review.
