### 2019-07-17 71f566e
Apply minor format corrections to 'paper.md'
