### 2019-07-17 b33b871
Add correct link to example figure in 'paper.md'
