### 2019-07-17 fda8dcf
Try to specify image width in paper.md
