+++ paper.md
@@ -110,7 +110,8 @@ fig = beam.plot()
 
 ### Output
 The script above produces the following Matplotlib figure:
-![Output corresponding to the example code above](https://github.com/alfredocarella/simplebendingpractice/raw/master/examples/example_1.png)
+
+![Output corresponding to the example code above](https://github.com/alfredocarella/simplebendingpractice/raw/master/examples/example_1.png =1000x)
 
 
 # Recent Uses
