### 2019-07-17 2956019
Remove figure width specification in paper.md
