+++ paper.md
@@ -111,7 +111,7 @@ fig = beam.plot()
 ### Output
 The script above produces the following Matplotlib figure:
 
-![Output corresponding to the example code above](https://github.com/alfredocarella/simplebendingpractice/raw/master/examples/example_1.png =1000x)
+![Output corresponding to the example code above](https://github.com/alfredocarella/simplebendingpractice/raw/master/examples/example_1.png)
 
 
 # Recent Uses
