### 2019-07-17 #60 @alfredocarella
Hi there! Thanks for starting the review process so promptly.
The paper includes a stubborn figure that keeps being placed at the end; sorry about that.
