### 2019-07-19 #60 @moorepants
@alfredocarella To fit as an "open-source learning module" you need to fit this criteria "Examples include Jupyter notebooks or plaintext/markup language documents like LaTeX, R Markdown, and ReST for course/lesson content and associated notes, with embedded or associated code snippets/programs." I'm not finding the documents + associated notes in your repository.
