### 2019-07-19 #60 @alfredocarella
@moorepants The lesson content is a chapter of the documentation (3 consecutive pages including code snippets) starting here: https://alfredocarella.github.io/simplebendingpractice/background.html
After going through that content, the concepts can be applied starting from this example: https://alfredocarella.github.io/simplebendingpractice/examples/example_1.html
which is also provided in the form of an online hosted Jupyter notebook.
