### 2019-07-19 #60 @moorepants
Ok, that's what I was looking for. Thank you. I think this meets the basic criteria, I'll round up some reviewers.
