### 2019-07-22 #60 @moorepants
@faradai2008 or @nagordon, would either of you be interested in reviewing this article submitted to the Journal of Open Source Education on a package for teaching about beam bending mechanics?
