### 2019-07-25 #60 @moorepants
Destiny Garcia (@desgarcia) from UC Davis has agreed to be a reviewer. I'm still hoping to get one more additional reviewer.

Destiny, you can begin reviewing the submitted paper, software repository, and software documentation from the links above, but the review checklist will be created in a new github issue once I get the second reviewer (here is an example you can check out to see how it works: https://github.com/openjournals/jose-reviews/issues/44)
