### 2019-07-26 #60 @moorepants
@buddyd16 Might you be interested in being a reviewer for this submission to the Journal of Open Source Education? I see that you have written a Python based structural engineering package that gives you unique experience for reviewing this similar software and materials.
