### 2019-07-26 #60 @nicoguaro
@moorepants I can review it.
