### 2019-07-26 #60 @moorepants
Fantastic! @nicoguaro Please read the reviewer guidelines here: http://jose.theoj.org/about#reviewer_guidelines and both reviewers should note that this is an "open-source educational materials/learning module" submission.
