### 2019-07-26 #60 @moorepants
Now that we have two reviewers I will open the actual review issue with the checklist. @nicoguaro and @desgarcia, you can review at your own pace and bring up any discussions on the new issue. I would like to target two weeks for the reviews to be completed (by August 9th). Let me know if that will be an issue. Thank you again for your time and expertise in reviewing.
