### 2019-07-26 #65 @whedon
[REVIEW]: BeamBending: a teaching aid for 1-D shear-force and bending-moment diagrams
**Submitting author:** @alfredocarella (<a href="http://orcid.org/0000-0001-8340-3427">Alfredo Carella</a>)
**Repository:** <a href="https://github.com/alfredocarella/simplebendingpractice" target ="_blank">https://github.com/alfredocarella/simplebendingpractice</a>
**Version:** v1.0.1
