### 2019-07-26 #60 @moorepants
Closing, if any other reviewers would like to join please respond in issue #65.
