### 2019-07-26 #65 @nicoguaro
@moorepants, should the repository include two license files?

@alfredocarella, do you have a DOI for your reference? Also, I would suggest another reference that is not in Norwegian, preferably in English.
