### 2019-07-26 #65 @moorepants
> @moorepants, should the repository include two license files?

I'm personally fine with a single license for all materials in the repository. Some people license code separately than other content (text, images, etc.). That is fine too.
