### 2019-07-29 be4c988
Add reference in English and JOSE status badge.
