+++ paper.bib
@@ -13,7 +13,7 @@
   year      = {2017},
   author    = {Beer, Ferdinand and Russell Johnston, E. and Dewolf, John T. and Mazurek, David},
   edition   = {Second},
-  isbn      = {978-0073398167},
+  isbn      = {978-0-07-339816-7},
 }
 
 @Comment{jabref-meta: databaseType:bibtex;}
