### 2019-07-29 #65 @alfredocarella
> @alfredocarella, do you have a DOI for your reference? Also, I would suggest another reference that is not in Norwegian, preferably in English.

@nicoguaro Sorry about the late answer.
I have unfortunately not found any DOI for Bell's book, but its ISBN (978-82-450-1685-7) is given in the file `paper.bib`.
As you very well point out, it is probably a good idea to have a reference in English. I have therefore added "Statics and Mechanics of Materials" by F. Beer (ISBN 978-0-07-339816-7). The relevant sections (i.e. where the topic is introduced) are 12.1 and 12.2.
