### 2019-07-29 2267f7a
Minor format corrections to paper references.
