+++ paper.bib
@@ -12,7 +12,7 @@
   publisher = {{McGraw-Hill Education}},
   year      = {2017},
   author    = {Beer, Ferdinand and Russell Johnston, E. and Dewolf, John T. and Mazurek, David},
-  edition   = {Second},
+  edition   = {Second Edition},
   isbn      = {978-0-07-339816-7},
 }
 
