### 2019-07-29 #60 @alfredocarella
I am sending the pdf generation command here in order to avoid polluting the other issue with useless output.
