### 2019-07-29 #65 @nicoguaro
@alfredocarella, it seems that the package only supports statically determined beams with (exactly) one pinned and one roller support. Maybe it is a good idea to explicitly mention this in the documentation and in the paper.

It seems a little bit confusing to me to use interchangeably pinned and fixed. Since the support doesn't constrain rotations.
