### 2019-07-31 595ed08
Correct the term "fixed support" to "pinned support" in code and documentation.
