+++ README.md
@@ -46,10 +46,10 @@ from beambending import Beam
 beam = Beam(9)  # Initialize a Beam object of length 9m
 ```
 
-After a `Beam` object is created, the properties corresponding to the x-coordinates of the fixed and rolling supports must be defined.
+After a `Beam` object is created, the properties corresponding to the x-coordinates of the pinned and rolling supports must be defined.
 
 ```python
-beam.fixed_support = 2    # x-coordinate of the fixed support
+beam.pinned_support = 2    # x-coordinate of the pinned support
 beam.rolling_support = 7  # x-coordinate of the rolling support
 ```
 
