### 2019-07-31 8cc7c8a
Do minor corrections to documentation and paper.
