### 2019-07-31 #65 @alfredocarella
@nicoguaro, thanks for two useful comments. I have done the following to address them:
* The mentioned limitations in package scope (i.e. only statically determined beams with one pinned and one roller support) is now explicitly mentioned in the documentation, class docstrings, `paper.md`, and `README.md`.
* I have also changed all code and documentation references from "fixed support" to "pinned support", in hopes of making it clearer.
