### 2019-07-31 #65 @nicoguaro
@moorepants, I have some questions regarding the "common" practices in JOSE.

- How specific should the community guidelines be?

- Is it OK to suggest to use vector (instead of raster) graphics in the documentation? This would make easier to repurpose the lessons, and even let to automate the translation of them.
