### 2019-07-31 #65 @nicoguaro
@alfredocarella, I opened [issue #1](https://github.com/alfredocarella/simplebendingpractice/issues/1) regarding the tests.

I also have some comments regarding units, that might be important considering that this is educational content:

- Units should be preceded by a space (see [SI Brochure](http://www.bipm.org/en/si/si_brochure/), [§5.3.3](http://www.bipm.org/en/publications/si-brochure/section5-3.html)). This is done sometimes in the package/documentation and sometimes is not.

- The package assume a non-SI system of units (m, kN, kN·m). I think that a (brief) mention of the units would be useful.

- Units should be roman in roman type and not italic (see [this](https://physics.nist.gov/cuu/Units/checklist.html)). This is sometimes taken into account in the package/documentation and sometimes not.
