### 2019-07-31 #65 @moorepants
> How specific should the community guidelines be?

One way to handle this is to add a `CONTRIBUTING` file to the repository. It does not need to be elaborate and just needs to address the items in the checklist, even if is says that contributions aren't accepted.

> Is it OK to suggest to use vector (instead of raster) graphics in the documentation? This would make easier to repurpose the lessons, and even let to automate the translation of them.

It is fine to suggest this. It will depend on the author's response as to whether it is possible.
