### 2019-07-31 alfredocarella/simplebendingpractice#2 @nicoguaro
Version badge points to a dead-end
The badge points to the following link:

     https://github.com/alfredocarella/simplebendingpractice/blob/master/None

Maybe the author is waiting after the [JOSE Review](https://github.com/openjournals/jose-reviews/issues/65) to create a release. In that case, this Issue would be closed accordingly.
