### 2019-08-03 #65 @moorepants
@alfredocarella Can you please cite the primary software dependencies in the paper? Additionally, there are examples of existing software with similar capabilities that are not mentioned in your paper. I'd like to see citations to at least some of these and some note of what your work offers that distinguishes it from them.
