### 2019-08-05 #65 @moorepants
@desgarcia and @nicoguaro I see pretty solid progress in the checklists above. This is a reminder that we'd like to complete the reviews by this Friday if possible. Let me know if you have any questions or need help in any way. Can each of you respond here with a check in letting me know if you think you'll be able to finish by the end of Friday? Thank you for your time and expertise.
