### 2019-08-05 #65 @nicoguaro
@moorepants, if my comments (above and Issue tracker) are addressed I would be able.
