### 2019-08-08 #65 @alfredocarella
@nicoguaro @desgarcia @moorepants Thanks for the feedback. I will start addressing all your suggestions as soon as possible.
