### 2019-08-09 b944be1
Fix numerical evaluation of symbolic functions
The previous code breaks with sympy > 1.1.
The implemented workaround evaluates one point at the time instead of using vectors.
