### 2019-08-09 6b19881
Change Python version to 3.7 and sympy version to latest (1.4 as of today)
