+++ requirements.txt
@@ -1,8 +1,8 @@
 matplotlib  # for plotting
 numpy  # for numerical calculations
 numpydoc  # numpy-style docstrings for sphinx
-pytest>=3.7  # for testing
+pytest>=4.1  # for testing
 pytest-cov  # for calculating coverage
 sphinx  # generating documentation
 sphinx_rtd_theme  # docs theme
-sympy==1.1  # for analytical integration
+sympy  # for analytical integration
