+++ .travis.yml
@@ -1,7 +1,8 @@
 # ref: https://docs.travis-ci.com/user/languages/python/
+dist: xenial   # required for Python >= 3.7
 language: python
 python:
-  - "3.6"
+  - "3.7"
 
 # ref: https://docs.travis-ci.com/user/customizing-the-build/#building-specific-branches
 branches:
