### 2019-08-09 a010ca7
Update recipe for using "xvfb" with "dist:xenial" and "python:3.7"
