+++ .travis.yml
@@ -1,5 +1,7 @@
 # ref: https://docs.travis-ci.com/user/languages/python/
 dist: xenial   # required for Python >= 3.7
+services:
+  - xvfb
 language: python
 python:
   - "3.7"
@@ -16,10 +18,10 @@ install:
 
 # set up the display port to generate figures
 # ref: https://docs.travis-ci.com/user/gui-and-headless-browsers/#using-xvfb-directly
-before_script:
-  - "export DISPLAY=:99.0"
-  - "sh -e /etc/init.d/xvfb start"
-  - sleep 3 # give xvfb some time to start
+# before_script:
+#   - "export DISPLAY=:99.0"
+#   - "sh -e /etc/init.d/xvfb start"
+#   - sleep 3 # give xvfb some time to start
 
 # command to test examples and then run tests
 script:
