### 2019-08-09 alfredocarella/simplebendingpractice#1 @alfredocarella
@nicoguaro I see the package did not work with sympy>1.1.
It looks like it works fine now.
