### 2019-08-09 alfredocarella/simplebendingpractice#2 @alfredocarella
>
>
> The badge points to the following link:
>
> ```
>  https://github.com/alfredocarella/simplebendingpractice/blob/master/None
> ```
>
> Maybe the author is waiting after the [JOSE Review](https://github.com/openjournals/jose-reviews/issues/65) to create a release. In that case, this Issue would be closed accordingly.

Precisely. I would like to at least correct all the obvious mistakes and release a working version.
