### 2019-08-09 alfredocarella/simplebendingpractice#1 @nicoguaro
They pass now.
