### 2019-08-09 d9c0049
Specify guidelines for contributing in a separate file "CONTRIBUTING.md"
