### 2019-08-10 #65 @moorepants
@nicoguaro and @desgarcia Thank you very much for the timely reviews. I think that everything looks reasonable and that minor revisions are needed for publication. @alfredocarella Please address any issues posted on your repository, issues in this thread (including missing checkboxes in the list above). These are the main things I'm seeing:

- Adding/improving community contribution guidelines
- Improving the explanation of units
- Citing prior art
- Improving installation and getting started aspects
- Student feedback (if available)

Feel free to discuss with the reviewers and me anything in as you make your revision. Will an August 23rd date be suitable to finish the revisions by?
