### 2019-08-23 5e3e4be
Update package requirements (remove restriction on pytest)
