+++ requirements.txt
@@ -1,7 +1,7 @@
 matplotlib  # for plotting
 numpy  # for numerical calculations
 numpydoc  # numpy-style docstrings for sphinx
-pytest>=4.1  # for testing
+pytest  # for testing
 pytest-cov  # for calculating coverage
 sphinx  # generating documentation
 sphinx_rtd_theme  # docs theme
