### 2019-08-23 95e9bcd
Change link to paid applet for a free Khan Academy video
