### 2019-08-23 b0bcc12
Correct how units are displayed
Preceed units by a space in theory module and example
Remove italics
