### 2019-08-23 83c797f
Add comparison with other software tools to paper.md
