### 2019-08-23 63576fe
Upload source for comparison table in paper.md
