### 2019-08-23 #65 @alfredocarella
@moorepants @nicoguaro @desgarcia
> @nicoguaro and @desgarcia Thank you very much for the timely reviews. I think that everything looks reasonable and that minor revisions are needed for publication. @alfredocarella Please address any issues posted on your repository, issues in this thread (including missing checkboxes in the list above). These are the main things I'm seeing:
>
>     * Adding/improving community contribution guidelines
>
>     * Improving the explanation of units
>
>     * Citing prior art
>
>     * Improving installation and getting started aspects
>
>     * Student feedback (if available)
>

* The (short) file ```CONTRIBUTING.md``` was added to the repository, in order to explicitly state that contributions are welcome. A small project like this will probably not benefit from having long and complex guidelines.
* The units in both documentation and generated plots are now in roman type, and preceded by a space. A brief comment has been added to the main class documentation, explicitly stating the (non-SI) units used by the package.
* The "Statement of Need" section of the paper includes now a brief comparison with similar available software (summarized in a table), including the main points that distinguish "beambending" from the rest.
* A link has been added to a tutorial on how to install Python 3 on either Linux, Windows or MacOS. Installation of the package can even be skipped altogether by just using the provided Jupyter notebook hosted on mybinder.
* No proper student feedback is available. The comments I received were oral, informal and unspecific. Real concrete examples are all along the lines of:
"I found it easy to check my answers to the practice problems", or "Playing around with the load shapes helped me visualize how distributed loads, shear force and bending moment are correlated"
However, I have no objective reason to assume that this feedback is representative or complete, since:
a) I did not receive a lot of comments (about a dozen, which is roughly 10 percent of the class), and
b) the comments are likely to be biased in a positive directions (few students spontaneously approach a teacher to tell him that his pet-project is terrible).

Additionally:
* Vector (instead of raster) graphics sounds like a good suggestion. I actually hoped to be able to do this in the first version. However, I found it quite time consuming to generate the vector figures. As of today the documentation consists of 14 figures. Five (5) of these are generated automatically by the package. Converting these to vector format is just a matter of replacing ".png" by (for example) ".svg" in the file name passed to the ```savefig``` function. The remaining 9 figures have been manually edited and would require some extra effort, but this should be manageable for the next update.
