### 2019-08-23 16df918
Add references to main python packages
