### 2019-08-23 61ee6ed
Add references to similar software
