+++ paper.bib
@@ -56,4 +56,52 @@
  doi = {10.7717/peerj-cs.103}
 }
 
+@misc{beamguru,
+  title = {{Beam Guru}},
+  howpublished = {\url{http://beamguru.com}},
+  note = {Accessed: 2019-08-22}
+}
+
+@misc{structuralbeam,
+  title = {{Structural Beam Deflection and Stress Calculators}},
+  howpublished = {\url{https://www.amesweb.info/StructuralBeamDeflection/BeamDeflectionCalculators.aspx}},
+  note = {Accessed: 2019-08-22}
+}
+
+@misc{skyciv,
+  title = {{SkyCiv Beam}},
+  howpublished = {\url{https://skyciv.com/structural-software/beam-analysis-software/}},
+  note = {Accessed: 2019-08-22}
+}
+
+@misc{mechanicalc,
+  title = {{MechaniCalc}},
+  howpublished = {\url{https://mechanicalc.com/calculators/beam-analysis/}},
+  note = {Accessed: 2019-08-22}
+}
+
+@misc{engineersedge,
+  title = {{Engineers Edge}},
+  howpublished = {\url{https://www.engineersedge.com/}},
+  note = {Accessed: 2019-08-22}
+}
+
+@misc{webstructural,
+  title = {{WebStructural}},
+  howpublished = {\url{https://webstructural.com/shear-and-moment-diagram.html}},
+  note = {Accessed: 2019-08-22}
+}
+
+@misc{beamcalculatoronline,
+  title = {{Beam Calculator Online}},
+  howpublished = {\url{http://rascheta.net/beamuk/}},
+  note = {Accessed: 2019-08-22}
+}
+
+@misc{steelbeamcalculator,
+  title = {{Steel Beam Calculator}},
+  howpublished = {\url{https://www.steelbeamcalculator.com/}},
+  note = {Accessed: 2019-08-22}
+}
+
 @Comment{jabref-meta: databaseType:bibtex;}
