### 2019-08-23 #65 @nicoguaro
Given the new inclusions to the paper, I have the following comments:

- I suggest adding bullet points/numerals to the second and to the last paragraphs on "Statement of Need".

- Table 1 is currently a figure instead of text. Is there a way to include this as text in the document?

- The inclusion of Table 1 provides a comparison of features with "similar" software. Nevertheless, it might be useful to add "bibliographic" information to those packages.
