### 2019-08-23 8b8c5a9
Add references to the text for main packages and software tools
