### 2019-08-24 b2b7ec3
Add forgotten spaces before bullet lists
