+++ paper.md
@@ -44,8 +44,9 @@ Introducing a topic that resembles the mindset of calculus and employs a little 
 <!-- This becomes an additional challenge for students in their first encounter with the topic of shear forces and bending moments in beams. -->
 
 This package/module aims to bridge this gap and simplify students' first contact with this challenging new topic by working on two fronts simultaneously:
-1. Explain the [background theory](https://alfredocarella.github.io/simplebendingpractice/background.html) from a simple example with focus on connecting the mathematical description with the physical beam model (`beambending` code snippets are interleaved in order to illustrate how the package works).
-2. Provide a temporary scaffolding that helps to establish an immediate visual association between beam load states and internal stresses.
+
+* Explain the [background theory](https://alfredocarella.github.io/simplebendingpractice/background.html) from a simple example with focus on connecting the mathematical description with the physical beam model (`beambending` code snippets are interleaved in order to illustrate how the package works).
+* Provide a temporary scaffolding that helps to establish an immediate visual association between beam load states and internal stresses.
 
 Several online tools with similar functionality are currently available, such as [@beamguru], [@structuralbeam], [@skyciv], [@mechanicalc], [@engineersedge], [@webstructural], [@beamcalculatoronline], and [@steelbeamcalculator].
 Most of them expose only a graphical user interface to the user, eliminating the need to write any code; roughly half of the surveyed tools are free while others charge monthly subscriptions; and only a few include a theoretical module or present to the user a detailed solution procedure for the problems.
@@ -54,6 +55,7 @@ This feature comparison is presented in the Table 1.
 ![](https://github.com/alfredocarella/simplebendingpractice/raw/master/tool_comparison_table.png)
 
 There are three main differences between the ```beambending``` package and the rest of the reviewed tools:
+
 * Arbitrary distributed load functions are accepted (as long as sympy parse them), i.e. they are not restricted to constants or linear functions.
 * The package (and parts of it) can be called from regular Python code, which makes it easy to automatically generate problem sets with solutions.
 * It is not only free, but also completely open source.
