### 2019-08-24 #65 @nicoguaro
@alfredocarella
>
>     * I would really like to add that table as text, but I have not been able to either embed latex in the markdown (the source code for the table is in the repository), or make that kind of table using markdown (which really results illegible without rotating the first row since the text is too long). Pasting a figure was a workaround, but I am willing to correct it if any of you can provide indication.

Can you share the LaTeX code? Maybe I can provide some ideas.

>     * The size and position of the figures is not great, but I have unfortunately not been able to do much better than that.

If you use SVG instead of PNG, does whedon compile the paper properly?
