### 2019-08-24 #65 @alfredocarella
> Can you share the LaTeX code? Maybe I can provide some ideas.
Here is the link to the LaTeX code: https://github.com/alfredocarella/simplebendingpractice/raw/master/tool_comparison_table.tex

> If you use SVG instead of PNG, does whedon compile the paper properly?

I am about to try exactly that... in a minute
