### 2019-08-24 61066ea
Change format for example_1 from "png" to "svg"
