### 2019-08-24 15e0b06
Deploy alfredocarella/simplebendingpractice to github.com/alfredocarella/simplebendingpractice.git:gh-pages
