### 2019-08-24 cd9aa74
Revert "Change format for example_1 from "png" to "svg""
This reverts commit 61066eac4daca5b7bb03c410fb182ee69aa0384a.
