### 2019-08-28 152e152
Minor format changes to paper and package
* Change size and position of figure on paper
* Move "accessed on DATE" information for references to a different field on .bib file
* Point to description-file README.md on setup.cfg
* Add some more metadata to setup.py
