+++ setup.py
@@ -10,7 +10,7 @@ from setuptools import setup
 
 
 setup(name='beambending',  # name of package on import
-      version='1.0',  # package version
+      version='1.0.0',  # package version
       description='Educational package for visualizing shear forces and bending moments on 1-D beams',  # brief description
       license='cc-by-4.0',  # licensing
       long_description=open('README.md').read(),
@@ -28,4 +28,10 @@ setup(name='beambending',  # name of package on import
         'sphinx_rtd_theme',  # docs theme
         'sympy'  # for analytical integration
       ],
-      zip_safe=True)  # package can be installed from zip file
+    classifiers=[
+      'Development Status :: 4 - Beta',         # Chose either "3 - Alpha", "4 - Beta" or "5 - Production/Stable" as the current state of your package    
+      'License :: OSI Approved :: cc-by-4.0',   # Again, pick a license
+      'Programming Language :: Python :: 3',    #Specify which pyhton versions that you want to support
+    ],
+      zip_safe=True  # package can be installed from zip file
+)
