+++ setup.cfg
@@ -1,3 +1,6 @@
+# point to the file describing the package
+[metadata]
+description-file = README.md
 # define explicitly the test files pytest will collect
 [tool:pytest]
 python_files = test_*.py *_test.py test/*.py example_*.py
