### 2019-08-28 0e94ab6
Make another attempt to fix paper format
