+++ paper.bib
@@ -59,49 +59,49 @@
 @misc{beamguru,
   title = {{Beam Guru}},
   howpublished = {\url{http://beamguru.com}},
-  year = {Accessed on 2019-08-22}
+  publisher = {{Accessed on 2019-08-22}}
 }
 
 @misc{structuralbeam,
   title = {{Structural Beam Deflection and Stress Calculators}},
   howpublished = {\url{https://www.amesweb.info/StructuralBeamDeflection/BeamDeflectionCalculators.aspx}},
-  year = {Accessed on 2019-08-22}
+  publisher = {{Accessed on 2019-08-22}}
 }
 
 @misc{skyciv,
   title = {{SkyCiv Beam}},
   howpublished = {\url{https://skyciv.com/structural-software/beam-analysis-software/}},
-  year = {Accessed on 2019-08-22}
+  publisher = {{Accessed on 2019-08-22}}
 }
 
 @misc{mechanicalc,
   title = {{MechaniCalc}},
   howpublished = {\url{https://mechanicalc.com/calculators/beam-analysis/}},
-  year = {Accessed on 2019-08-22}
+  publisher = {{Accessed on 2019-08-22}}
 }
 
 @misc{engineersedge,
   title = {{Engineers Edge}},
   howpublished = {\url{https://www.engineersedge.com/}},
-  year = {Accessed on 2019-08-22}
+  publisher = {{Accessed on 2019-08-22}}
 }
 
 @misc{webstructural,
   title = {{WebStructural}},
   howpublished = {\url{https://webstructural.com/shear-and-moment-diagram.html}},
-  year = {Accessed on 2019-08-22}
+  publisher = {{Accessed on 2019-08-22}}
 }
 
 @misc{beamcalculatoronline,
   title = {{Beam Calculator Online}},
   howpublished = {\url{http://rascheta.net/beamuk/}},
-  year = {Accessed on 2019-08-22}
+  publisher = {{Accessed on 2019-08-22}}
 }
 
 @misc{steelbeamcalculator,
   title = {{Steel Beam Calculator}},
   howpublished = {\url{https://www.steelbeamcalculator.com/}},
-  year = {Accessed on 2019-08-22}
+  publisher = {{Accessed on 2019-08-22}}
 }
 
 @Comment{jabref-meta: databaseType:bibtex;}
