### 2019-08-28 77b6ffd
And yet another attempt to get the paper formatting right
