### 2019-08-28 #65 @alfredocarella
@moorepants @nicoguaro @desgarcia
Sorry about polluting the main issue with all the paper formatting attempts.
I believe all the issues and comments have been addressed, and the paper is ready for a second review.
