### 2019-08-28 9c9b3b2
Grammar check on the paper
