+++ paper.md
@@ -20,10 +20,10 @@ bibliography: paper.bib
 
 # Summary
 BeamBending is _both_ an educational module _and_ a Python package, based mainly on [@Hunter:2007], [@oliphant2006guide] and [@10.7717/peerj-cs.103] It is intended to serve as a teaching aid during a first course in Statics.
-The aim of this module is to enhance clarity and provide visual hands-on examples while introducing the concepts of:
+This module aims to enhance clarity and provide visual hands-on examples while introducing the concepts of:
 
 * stresses on slender _one-dimensional_ solids (i.e. beams)
-* normal force, shear force and bending moment diagrams
+* normal force, shear force, and bending moment diagrams
 
 The [package documentation](https://alfredocarella.github.io/simplebendingpractice/) includes a simple (but still rigorous enough) explanation of the background theory, inspired in [@Beer2017] and [@Bell2015].
 It is assumed that the students understand static equilibrium of flat rigid bodies, but a short recap is provided.
@@ -35,7 +35,7 @@ The package can be used by
 * students who want to verify their solutions to introductory problem sets;
 * students who like to play with example problems and receive immediate visual feedback (i.e. about how simple modifications to imposed loads affect the resulting reaction forces and internal stresses).
 
-The `beambending` package is ready for installation using `pip`, or can be tested online using the provided [Jupyter notebook](https://mybinder.org/v2/gh/alfredocarella/simplebendingpractice/master?filepath=simple_demo.ipynb).
+The `beambending` package is ready for installation using `pip` or can be tested online using the provided [Jupyter notebook](https://mybinder.org/v2/gh/alfredocarella/simplebendingpractice/master?filepath=simple_demo.ipynb).
 
 
 # Statement of Need
@@ -45,12 +45,12 @@ Introducing a topic that resembles the mindset of calculus and employs a little 
 
 This package/module aims to bridge this gap and simplify students' first contact with this challenging new topic by working on two fronts simultaneously:
 
-* Explain the [background theory](https://alfredocarella.github.io/simplebendingpractice/background.html) from a simple example with focus on connecting the mathematical description with the physical beam model (`beambending` code snippets are interleaved in order to illustrate how the package works).
+* Explain the [background theory](https://alfredocarella.github.io/simplebendingpractice/background.html) from a simple example with focus on connecting the mathematical description with the physical beam model (`beambending` code snippets are interleaved to illustrate how the package works).
 * Provide a temporary scaffolding that helps to establish an immediate visual association between beam load states and internal stresses.
 
 Several online tools with similar functionality are currently available, such as [@beamguru], [@structuralbeam], [@skyciv], [@mechanicalc], [@engineersedge], [@webstructural], [@beamcalculatoronline], and [@steelbeamcalculator].
-Most of them expose only a graphical user interface to the user, eliminating the need to write any code; roughly half of the surveyed tools are free while others charge monthly subscriptions; and only a few include a theoretical module or present to the user a detailed solution procedure for the problems.
-This feature comparison is presented in the Table 1.
+Most of them expose only a graphical user interface to the user, eliminating the need to write any code; roughly half of the surveyed tools are free while others charge monthly subscriptions, and only a few include a theoretical module or present to the user a detailed solution procedure for the problems.
+This feature comparison is presented in Table 1.
 
 ![](https://github.com/alfredocarella/simplebendingpractice/raw/master/tool_comparison_table.png)
 
@@ -58,7 +58,7 @@ There are three main differences between the ```beambending``` package and the r
 
 * Arbitrary distributed load functions are accepted (as long as sympy parse them), i.e. they are not restricted to constants or linear functions.
[…]
