### 2019-08-28 cb82f28
Add long_description_content_type to setup.py
