### 2019-08-28 0566432
Change installation instructions after having uploaded the package to PyPI
