### 2019-08-29 #65 @moorepants
> @moorepants, is there any workaround to include SVG in the paper?

I don't know specifically. It should support anything that pandoc markdown supports.
