### 2019-08-29 #65 @moorepants
>  I believe all the issues and comments have been addressed, and the paper is ready for a second review.

@alfredocarella Great! Thank you for addressing all of the comments.

@nicoguaro and @desgarcia Can you all comment on whether the changes are sufficient? If not, please make any final specific suggestions. If there are no major issues I will recommend for publishing.
