### 2019-08-29 #65 @moorepants
> Table 1 is currently a figure instead of text. Is there a way to include this as text in the document?

I'm fine with it as an image. IMHO, the markdown format is extremely limiting and it is painful to create many of the things that are relatively easiy with LaTeX.
