### 2019-08-29 #65 @nicoguaro
@moorepants, I consider that the paper is ready for publication.
