### 2019-08-29 #65 @nicoguaro
@moorepants

> I don't know specifically. It should support anything that pandoc markdown supports.

It does not support it when compiling to LaTeX. I asked because I couldn't find any issue related to this in whedon's repo.

> I'm fine with it as an image. IMHO, the markdown format is extremely limiting and it is painful to create many of the things that are relatively easiy with LaTeX.

I'm also fine with it. I prefer the other option and that's why I asked. I also agree on the comment about markdown.
