### 2019-08-29 #65 @moorepants
>It does not support it when compiling to LaTeX. I asked because I couldn't find any issue related to this in whedon's repo.

I'd avoid it then. Maybe open an issue on the Whedon repo with a feature request?

> I'm also fine with it. I prefer the other option and that's why I asked.

if you can help get it working, go for it, but we shouldn't let this hold things up too much.
