### 2019-08-29 #65 @desgarcia
@moorepants I also agree with @nicoguaro , I think the paper is ready for publication
