### 2019-08-29 #65 @moorepants
@alfredocarella I see one error in a phrase on page 2: "as long as sympy <can> parse them"
