### 2019-08-29 #65 @moorepants
After fixing that, can you tag a release, submit an archived version to zenodo, and then share the DOI here?
