### 2019-08-29 #65 @labarba
Please also check the citation syntax.

When the citation is part of the sentence, then use: Author (year). If the citation is not part of the sentence, then it should appear in parenthesis. E.g. “We take inspiration from Smith et al. (2006) to … “ / 	“The code implements Theory 1 (Jones 2012) and Theory 2 (Roberts 2014) to obtain … “

For the syntax to obtain brackets in the right places, see https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax
