### 2019-08-29 #65 @labarba
What is meant by "a little intuitive sign conventions"? Do you mean _not_ intuitive? Intuitive but just a little? Please reword.
