### 2019-08-30 dad994f
Correct grammar and citation format in paper
