+++ paper.bib
@@ -16,7 +16,7 @@
   isbn      = {978-0-07-339816-7},
 }
 
-@Article{Hunter:2007,
+@Article{Hunter2007,
   Author    = {Hunter, J. D.},
   Title     = {Matplotlib: A 2D graphics environment},
   Journal   = {Computing in Science \& Engineering},
@@ -39,7 +39,7 @@
   publisher={Trelgol Publishing USA}
 }
 
-@article{10.7717/peerj-cs.103,
+@article{sympy2017,
  title = {{SymPy: symbolic computing in Python}},
  author = {Meurer, Aaron and Smith, Christopher P. and Paprocki, Mateusz and \v{C}ert\'{i}k,
  Ond\v{r}ej and Kirpichev, Sergey B. and Rocklin, Matthew and Kumar, AMiT and Ivanov,
