### 2019-08-30 41934df
Write software citations by hand to correct their format
