### 2019-09-02 6897d12
Launch Release 1.0.1 with minor corrections to documentation
