### 2019-09-02 #65 @alfredocarella
>
>
> After fixing that, can you tag a release, submit an archived version to zenodo, and then share the DOI here?

@moorepants Done! The DOI is `10.5281/zenodo.3383748`
@labarba The suggested corrections have been made.
