### 2019-09-02 #65 @moorepants
One more minor fix. There is no date on your first reference in the bibliography, just empty parentheses.
