+++ paper.bib
@@ -110,7 +110,7 @@
 
 @misc{beamcalculatoronline,
   title = {{Beam Calculator Online}},
-  year = {},
+  year = {2019},
   howpublished = {\url{http://rascheta.net/beamuk/}},
   publisher = {{Accessed on 2019-08-22}}
 }
