### 2019-09-02 #65 @alfredocarella
>
>
> One more minor fix. There is no date on your first reference in the bibliography, just empty parentheses.

Sorry about that. Done.
