### 2019-09-02 #65 @moorepants
Ok, great!

@alfredocarella Congratulations, I recommend your paper for publication in JOSE!

@nicoguaro and @desgarcia Thank you both for your volunteered time and thoughtful reviews!

It was a pleasure working with everyone.

@labarba This one is ready for acceptance, if you could do the honors :smile:.
