### 2019-09-02 #65 @labarba
Big thanks to our editor: @moorepants, and the reviewers: @desgarcia, @nicoguaro — we couldn't do this without you! 🙏

Tweet announcing the publication:
https://twitter.com/JOSE_TheOJ/status/1168597259191275520
